fix: accept fractional share quantities in trade ingest
Robinhood fractional-share fills (e.g., "0.75 shares of SMH") were
failing the stock regex and landing in parse_failures, dropping the
fill silently. Widen the stock quantity capture to accept decimals,
parse as float, and switch quantity / net_qty schema declarations to
REAL so new installs have the right affinity. SQLite type affinity
preserves fractionals correctly in the existing INTEGER columns, so
no migration is needed for current databases.

+++ tools/trade_positions.py
@@ -90,7 +90,7 @@ def _ensure_schema(conn):
             strike REAL,
             option_type TEXT,
             expiration_date DATE,
-            net_qty INTEGER NOT NULL,
+            net_qty REAL NOT NULL,
             avg_buy_price REAL,
             total_cost REAL,
             opened_at DATETIME,
@@ -633,9 +633,9 @@ def _aggregate_executions(conn, position_key):
         notes_concat = "\n".join(lines)
 
     return {
-        'net_qty': int(row[0]),
+        'net_qty': float(row[0]),
         'buy_total_value': row[1] or 0.0,
-        'buy_total_qty': int(row[2] or 0),
+        'buy_total_qty': float(row[2] or 0),
         'net_total_cost': row[3],
         'opened_at': row[4],
         'last_action_at': row[5],
